requirements..txt
Pillow == 8.2.0
PyYAML == 5.4.1
anytree == 2.8.0
numpy == 1.21.6
opencv-python == 4.8.1.78
scikit-learn == 1.0.2
torch == 1.8.1
torchnet ==  0.0.4
torchvision == 0.9.1
tqdm == 4.50.2